connect sadeepa's page CancelJobPage.jsx,ImageUpload.jsx,Payment.jsx page
